+++ flutter/lib/src/event_processor/native_app_start_event_processor.dart
@@ -22,13 +22,26 @@ class NativeAppStartEventProcessor implements EventProcessor {
 
   @override
   Future<SentryEvent?> apply(SentryEvent event, Hint hint) async {
-    if (_native.didAddAppStartMeasurement || event is! SentryTransaction) {
+    final options = _hub.options;
+    if (_native.didAddAppStartMeasurement ||
+        event is! SentryTransaction ||
+        options is! SentryFlutterOptions) {
       return event;
     }
 
     final appStartInfo = await NativeAppStartIntegration.getAppStartInfo();
-    final measurement = appStartInfo?.toMeasurement();
 
+    final appStartEnd = _native.appStartEnd;
+    if (!options.autoAppStart) {
+      if (appStartEnd != null) {
+        appStartInfo?.end = appStartEnd;
+      } else {
+        // If autoAppStart is disabled and appStartEnd is not set, we can't add app starts
+        return event;
+      }
+    }
+
+    final measurement = appStartInfo?.toMeasurement();
     if (measurement != null) {
       event.measurements[measurement.name] = measurement;
       _native.didAddAppStartMeasurement = true;
@@ -44,6 +57,10 @@ class NativeAppStartEventProcessor implements EventProcessor {
   Future<void> _attachAppStartSpans(
       AppStartInfo appStartInfo, SentryTracer transaction) async {
     final transactionTraceId = transaction.context.traceId;
+    final appStartEnd = appStartInfo.end;
+    if (appStartEnd == null) {
+      return;
+    }
 
     final appStartSpan = await _createAndFinishSpan(
         tracer: transaction,
@@ -52,7 +69,9 @@ class NativeAppStartEventProcessor implements EventProcessor {
         parentSpanId: transaction.context.spanId,
         traceId: transactionTraceId,
         startTimestamp: appStartInfo.start,
-        endTimestamp: appStartInfo.end);
+        endTimestamp: appStartEnd);
+
+    await _attachNativeSpans(appStartInfo, transaction, appStartSpan);
 
     final pluginRegistrationSpan = await _createAndFinishSpan(
         tracer: transaction,
@@ -63,14 +82,14 @@ class NativeAppStartEventProcessor implements EventProcessor {
         startTimestamp: appStartInfo.start,
         endTimestamp: appStartInfo.pluginRegistration);
 
-    final mainIsolateSetupSpan = await _createAndFinishSpan(
+    final sentrySetupSpan = await _createAndFinishSpan(
         tracer: transaction,
         operation: appStartInfo.appStartTypeOperation,
-        description: appStartInfo.mainIsolateSetupDescription,
+        description: appStartInfo.sentrySetupDescription,
         parentSpanId: appStartSpan.context.spanId,
         traceId: transactionTraceId,
         startTimestamp: appStartInfo.pluginRegistration,
-        endTimestamp: appStartInfo.mainIsolateStart);
+        endTimestamp: appStartInfo.sentrySetupStart);
 
     final firstFrameRenderSpan = await _createAndFinishSpan(
         tracer: transaction,
@@ -78,15 +97,37 @@ class NativeAppStartEventProcessor implements EventProcessor {
         description: appStartInfo.firstFrameRenderDescription,
         parentSpanId: appStartSpan.context.spanId,
         traceId: transactionTraceId,
-        startTimestamp: appStartInfo.mainIsolateStart,
-        endTimestamp: appStartInfo.end);
+        startTimestamp: appStartInfo.sentrySetupStart,
+        endTimestamp: appStartEnd);
 
     transaction.children.addAll([
       appStartSpan,
       pluginRegistrationSpan,
-      mainIsolateSetupSpan,
+      sentrySetupSpan,
       firstFrameRenderSpan
     ]);
+  }
+
+  Future<void> _attachNativeSpans(AppStartInfo appStartInfo,
+      SentryTracer transaction, SentrySpan parent) async {
+    await Future.forEach<TimeSpan>(appStartInfo.nativeSpanTimes,
+        (timeSpan) async {
+      try {
+        final span = await _createAndFinishSpan(
+            tracer: transaction,
+            operation: appStartInfo.appStartTypeOperation,
+            description: timeSpan.description,
+            parentSpanId: parent.context.spanId,
+            traceId: transaction.context.traceId,
+            startTimestamp: timeSpan.start,
+            endTimestamp: timeSpan.end);
+        span.data.putIfAbsent('native', () => true);
+        transaction.children.add(span);
+      } catch (e) {
+        _hub.options.logger(SentryLevel.warning,
+            'Failed to attach native span to app start transaction: $e');
+      }
+    });
   }
 
   Future<SentrySpan> _createAndFinishSpan({