+++ flutter/lib/src/integrations/native_app_start_integration.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: invalid_use_of_internal_member
+
 import 'dart:async';
 
 import 'package:meta/meta.dart';
@@ -10,10 +12,16 @@ import '../event_processor/native_app_start_event_processor.dart';
 /// Integration which handles communication with native frameworks in order to
 /// enrich [SentryTransaction] objects with app start data for mobile vitals.
 class NativeAppStartIntegration extends Integration<SentryFlutterOptions> {
-  NativeAppStartIntegration(this._native, this._frameCallbackHandler);
+  NativeAppStartIntegration(this._native, this._frameCallbackHandler,
+      {Hub? hub})
+      : _hub = hub ?? HubAdapter();
 
   final SentryNative _native;
   final FrameCallbackHandler _frameCallbackHandler;
+  final Hub _hub;
+
+  /// Timeout duration to wait for the app start info to be fetched.
+  static const _timeoutDuration = Duration(seconds: 30);
 
   /// We filter out App starts more than 60s
   static const _maxAppStartMillis = 60000;
@@ -39,7 +47,8 @@ class NativeAppStartIntegration extends Integration<SentryFlutterOptions> {
     if (_appStartInfo != null) {
       return Future.value(_appStartInfo);
     }
-    return _appStartCompleter.future;
+    return _appStartCompleter.future
+        .timeout(_timeoutDuration, onTimeout: () => null);
   }
 
   @visibleForTesting
@@ -51,42 +60,48 @@ class NativeAppStartIntegration extends Integration<SentryFlutterOptions> {
   @override
   void call(Hub hub, SentryFlutterOptions options) {
     if (isIntegrationTest) {
-      final appStartInfo = AppStartInfo(AppStartType.cold,
-          start: DateTime.now(),
-          end: DateTime.now().add(const Duration(milliseconds: 100)),
-          pluginRegistration:
-              DateTime.now().add(const Duration(milliseconds: 50)),
-          mainIsolateStart:
-              DateTime.now().add(const Duration(milliseconds: 60)));
+      final appStartInfo = AppStartInfo(
+        AppStartType.cold,
+        start: DateTime.now(),
+        end: DateTime.now().add(const Duration(milliseconds: 100)),
+        pluginRegistration:
+            DateTime.now().add(const Duration(milliseconds: 50)),
+        sentrySetupStart: DateTime.now().add(const Duration(milliseconds: 60)),
+        nativeSpanTimes: [],
+      );
       setAppStartInfo(appStartInfo);
       return;
     }
 
-    if (options.autoAppStart) {
-      _frameCallbackHandler.addPostFrameCallback((timeStamp) async {
-        if (_native.didFetchAppStart) {
-          return;
-        }
+    if (_native.didFetchAppStart) {
+      return;
+    }
 
+    _frameCallbackHandler.addPostFrameCallback((timeStamp) async {
+      final nativeAppStart = await _native.fetchNativeAppStart();
+      if (nativeAppStart == null) {
+        setAppStartInfo(null);
+        return;
+      }
+
+      final sentrySetupStartDateTime = SentryFlutter.sentrySetupStartTime;
+      if (sentrySetupStartDateTime == null) {
+        setAppStartInfo(null);
+        return;
+      }
+
+      final appStartDateTime = DateTime.fromMillisecondsSinceEpoch(
+          nativeAppStart.appStartTime.toInt());
+      final pluginRegistrationDateTime = DateTime.fromMillisecondsSinceEpoch(
+          nativeAppStart.pluginRegistrationTime);
+      DateTime? appStartEndDateTime;
+
+      if (options.autoAppStart) {
         // We only assign the current time if it's not already set - this is useful in tests
-        // ignore: invalid_use_of_internal_member
         _native.appStartEnd ??= options.clock();
-        final appStartEndDateTime = _native.appStartEnd;
-        final nativeAppStart = await _native.fetchNativeAppStart();
-        final pluginRegistrationTime = nativeAppStart?.pluginRegistrationTime;
-        final mainIsolateStartDateTime = SentryFlutter.mainIsolateStartTime;
+        appStartEndDateTime = _native.appStartEnd;
 
-        if (nativeAppStart == null ||
-            appStartEndDateTime == null ||
-            pluginRegistrationTime == null) {
-          return;
-        }
-
-        final appStartDateTime = DateTime.fromMillisecondsSinceEpoch(
-            nativeAppStart.appStartTime.toInt());
-        final duration = appStartEndDateTime.difference(appStartDateTime);
-        final pluginRegistrationDateTime =
-            DateTime.fromMillisecondsSinceEpoch(pluginRegistrationTime);
+        final duration = appStartEndDateTime?.difference(appStartDateTime);
 
         // We filter out app start more than 60s.
         // This could be due to many different reasons.
@@ -96,23 +111,70 @@ class NativeAppStartIntegration extends Integration<SentryFlutterOptions> {
         // If the system forked the process earlier to accelerate the app start.
         // And some unknown reasons that could not be reproduced.
         // We've seen app starts with hours, days and even months.
-        if (duration.inMilliseconds > _maxAppStartMillis) {
+        if (duration != null && duration.inMilliseconds > _maxAppStartMillis) {
           setAppStartInfo(null);
           return;
         }
+      }
 
-        final appStartInfo = AppStartInfo(
-            nativeAppStart.isColdStart ? AppStartType.cold : AppStartType.warm,
-            start: appStartDateTime,
-            end: appStartEndDateTime,
-            pluginRegistration: pluginRegistrationDateTime,
-            mainIsolateStart: mainIsolateStartDateTime);
+      List<TimeSpan> nativeSpanTimes = [];
+      for (final entry in nativeAppStart.nativeSpanTimes.entries) {
+        try {
+          final startTimestampMs =
+              entry.value['startTimestampMsSinceEpoch'] as int;
+          final endTimestampMs =
+              entry.value['stopTimestampMsSinceEpoch'] as int;
+          nativeSpanTimes.add(TimeSpan(
+            start: DateTime.fromMillisecondsSinceEpoch(startTimestampMs),
+            end: DateTime.fromMillisecondsSinceEpoch(endTimestampMs),
+            description: entry.key as String,
+          ));
+        } catch (e) {
+          _hub.options.logger(
+              SentryLevel.warning, 'Failed to parse native span times: $e');
+          continue;
+        }
+      }
 
-        setAppStartInfo(appStartInfo);
-      });
-    }
+      // We want to sort because the native spans are not guaranteed to be in order.
+      // Performance wise this won't affect us since the native span amount is very low.
+      nativeSpanTimes.sort((a, b) => a.start.compareTo(b.start));
 
-    options.addEventProcessor(NativeAppStartEventProcessor(_native));
+      final appStartInfo = AppStartInfo(
+          nativeAppStart.isColdStart ? AppStartType.cold : AppStartType.warm,
+          start: appStartDateTime,
+          end: appStartEndDateTime,
+          pluginRegistration: pluginRegistrationDateTime,
+          sentrySetupStart: sentrySetupStartDateTime,
+          nativeSpanTimes: nativeSpanTimes);
+
+      setAppStartInfo(appStartInfo);
+
+      // When we don't have a SentryNavigatorObserver, a TTID transaction
+      // is not created therefore we need to create a transaction ourselves.
+      // We detect this by checking if the currentRouteName is null.
+      // This is a workaround since there is no api that tells us if
+      // the navigator observer exists and has been attached.
+      // The navigator observer also triggers much earlier so if it was attached
+      // it would have already set the routeName and the isCreated flag.
+      // The currentRouteName is always set during a didPush triggered
+      // by the navigator observer.
+      if (!SentryNavigatorObserver.isCreated &&
+          SentryNavigatorObserver.currentRouteName == null) {
+        const screenName = SentryNavigatorObserver.rootScreenName;
+        final transaction = hub.startTransaction(
+            screenName, SentrySpanOperations.uiLoad,
+            startTimestamp: appStartInfo.start);
+        final ttidSpan = transaction.startChild(
+            SentrySpanOperations.uiTimeToInitialDisplay,
+            description: '$screenName initial display',
+            startTimestamp: appStartInfo.start);
+        await ttidSpan.finish(endTimestamp: appStartInfo.end);
+        await transaction.finish(endTimestamp: appStartInfo.end);
+      }
+    });
+
+    options.addEventProcessor(NativeAppStartEventProcessor(_native, hub: hub));
 
     options.sdk.addIntegration('nativeAppStartIntegration');
   }
@@ -121,21 +183,33 @@ class NativeAppStartIntegration extends Integration<SentryFlutterOptions> {
 enum AppStartType { cold, warm }
 
 class AppStartInfo {
-  AppStartInfo(this.type,
-      {required this.start,
-      required this.end,
-      required this.pluginRegistration,
-      required this.mainIsolateStart});
+  AppStartInfo(
+    this.type, {
+    required this.start,
+    required this.pluginRegistration,
+    required this.sentrySetupStart,
+    required this.nativeSpanTimes,
+    this.end,
+  });
 
   final AppStartType type;
   final DateTime start;
-  final DateTime end;
+  final List<TimeSpan> nativeSpanTimes;
+
+  // We allow the end to be null, since it might be set at a later time
+  // with setAppStartEnd when autoAppStart is disabled
+  DateTime? end;
+
   final DateTime pluginRegistration;
-  final DateTime mainIsolateStart;
+  final DateTime sentrySetupStart;
 
-  Duration get duration => end.difference(start);
+  Duration? get duration => end?.difference(start);
 
-  SentryMeasurement toMeasurement() {
+  SentryMeasurement? toMeasurement() {
+    final duration = this.duration;
+    if (duration == null) {
+      return null;
+    }
     return type == AppStartType.cold
         ? SentryMeasurement.coldAppStart(duration)
         : SentryMeasurement.warmAppStart(duration);
@@ -146,6 +220,14 @@ class AppStartInfo {
   String get appStartTypeDescription =>
       type == AppStartType.cold ? 'Cold start' : 'Warm start';
   final pluginRegistrationDescription = 'App start to plugin registration';
-  final mainIsolateSetupDescription = 'Main isolate setup';
+  final sentrySetupDescription = 'Before Sentry Init Setup';
   final firstFrameRenderDescription = 'First frame render';
+}
+
+class TimeSpan {
+  TimeSpan({required this.start, required this.end, required this.description});
+
+  final DateTime start;
+  final DateTime end;
+  final String description;
 }